// generated/src/app/api-models/agent-commission.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AgentCommission {
  seqAgentCommissionId: number;
  seqAgentLicenseDetail: number;
  seqAgentLicenseHeader: number;
  agentSalesType: string;
  seqAgentId: number;
  commissionPercentage: number;
  seqVendId: number;
  seqVendAddress: number;
  userDefined1: string;
  userDefined2: string;
  userDefinedDate: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}